requirements.txt
openai==2.6.1
python-dotenv==1.1.1
Requests==2.32.5
